backend/requirements.txt
fastapi
uvicorn
pandas
xgboost
scikit-learn
shap
google-generativeai
python-dotenv
python-multipart
fastparquet
joblib
pymongo[srv]
motor